+++ cnccode/controller_classes/CTCNC.inc.php
@@ -237,6 +237,7 @@ class CTCNC extends Controller
         }
     }
 
+
     function hasPermissions($levels)
     {